100_Work/110_WorkApply/Project/VICI/requirements.txt
# LLM
google-genai>=1.0.0
# anthropic>=0.40.0

# PDF parsing
pymupdf>=1.24.0

# HTTP
requests>=2.32.0

# Environment
python-dotenv>=1.0.0
